Add an experimental prescaler named RAVU
RAVU is inspired by RAISR (Rapid and Accurate Image Super
Resolution). It adopts the core ideas of RAISR but did the
interpolating in a way similar to Super-xBR.

Don't use it, at least not in the current form. It's still a WIP,
and insanely slow at this moment.

The weights are trained with only a few random wallpaper/screenshots
files that I could find in local disk. They probably won't work well
for all kind of videos.