ravu: support model based on hermite polynomial
This should improve the PSNR a bit (as shown by result of regression
training). But I somehow doubt if it's really a good idea. Probably
I should also implement SSIM metric to evaluate the "improvement".

The weights are trained with the same set of pictures as linear
model, with the same parameters.