ravu: some workaround for r4-native distortion
It seems that putting 64 vec4s in function body could cause issue
with my driver. As a workaround, define all samples of native
profiles in global scope (they are static anyway).

I'm not sure if this actually solves the problem (though, the issue
is gone in my case). Anyway, this commit is marked to be reverted
later, if the issue persists.