ravu-zoom: address potential float precision issue
While I couldn't reproduce this issue using mpv, it has been observed in
the HLSL port of RAVU-Zoom for Magpie. Instead of digging and trying to
figure out how different shader compilers handle floats differently, it's
always easier just to stay on the safe side.

Thanks to @hauuau for report and testing work.

ref: Blinue/Magpie#803